1. 继续完善回复，自动链接书名.

+++ app/src/main/java/org/cryse/lkong/model/NewPostResult.java
@@ -14,7 +14,7 @@ public class NewPostResult {
         this.tid = tid;
     }
 
-    public long getPageCount() {
+    public int getPageCount() {
         return pageCount;
     }
